// src/app/(commonLayout)/_components/NewsFeed/Posts/posts.types.ts
export type TVote = {
    userId: string;
    votedAt: Date;
  };
  
  export type TComment = {
    authorId: string;
    commentedAt: Date;
    comment: string;
    likes: number;
  };
  
  export type TPost = {
    id: string;
    title: string;
    body: string;
    images?: string[];
    upvotes: TVote[];
    downvotes: TVote[];
    status?: 'draft' | 'published' | 'archived';
    contentType : "free" | "premium";
    comments: TComment[];
    category: string;
    createdAt: Date;
    authorId: string;
  };